% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scAgeCom_data}
\alias{scAgeCom_data}
\title{The scAgeCom data to be displayed on the ShinyApp}
\format{
A list
}
\usage{
data(scAgeCom_data)
}
\description{
The scAgeCom data to be displayed on the ShinyApp
}
\keyword{datasets}
